pulseaudio: rewrite output
this is a rewrite of the pulseaudio output. some major changes
include:

1) no more intermediate buffer
2) cleaner structure with all globals contained in a struct
3) ability to add more features in the future

One main problem is that the playback thread seems to disapear
when stopping playback. this can be a problem if we are waiting
for the pulse mainthread and the lock never being removed.
I've tried to cope with this by tracking the lock status but
IMO this should be handled better by the application main thread.