unveil(2) the few paths unless commands are run
In case nothing is to be executed at runtime, lock down filesystem access
down to
- write/create metadata pipe/cache
- read/rwite D-Bus socket
- read glib2 locales

If D-Bus setup were to happen earlier, we'd be able to ignore that
entirely (socket already opened) at this point and, iff metadata is off,
drop "wpath" as well, i.e. run with nothing but read-only locales.